requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
coverage==7.0.5
Django==4.1.5
django-cpf==0.1.0
django-dotenv==1.4.2
django-phonenumber-field==7.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.6.1
Flask==2.2.2
heyoo==0.0.8
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.3
phonenumbers==8.13.4
platformdirs==2.6.2
pluggy==1.0.0
pyrsistent==0.19.3
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-mock==3.10.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.14.0
